app/requirements.txt
Flask~=0.10
Flask-Script~=2.0
Flask-SQLAlchemy~=2.0
flask-cors~=2.1.2
Flask-Testing>=0.4.2
PyMySQL~=0.6
gunicorn~=19.3
requests>=0.11
